templates/stable-diffusion/requirements.txt
torch==2.2.0
diffusers==0.26.3
transformers==4.37.2
accelerate==0.26.1
safetensors==0.4.2
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pillow==10.2.0
xformers==0.0.24
